Change the date format from int to long

+++ app/src/main/java/com/leprincesylvain/ocproject7/go4lunch/model/Workmate.java
@@ -53,7 +53,7 @@ public class Workmate implements Parcelable {
         return restaurantId;
     }
 
-    public int getDateOfChoice() {
+    public long getDateOfChoice() {
         return dateOfChoice;
     }
